new: UsernamePasswordDTO, AuthenticationTokenDTO.
Created UsernamePasswordDTO that is used in the login method of class AdminController, and AuthenticationTokenDTO that is used
in the response body of the login method, it returns the token that the
admin can use to authenticates itself. Also changes in the security
layer, now if you provide a token that doesn't exists in a authenticated
request, you will get a customized response body.